Add endpoints for forcing vatsim data refresh Fixes #971
