+++ server/src/routes/airportInfo.mts
@@ -1,6 +1,4 @@
 import express from "express";
-import * as bree from "../bree.mjs";
-import { JobName } from "../bree.mjs";
 import { getAirportInfo } from "../controllers/airportInfo.mjs";
 import { verifyUser } from "../middleware/permissions.mjs";
 import { secureQueryMiddleware } from "../middleware/secureQueryMiddleware.mjs";
@@ -18,16 +16,6 @@ router.get("/airportInfo/:airportCode", verifyUser, secureQueryMiddleware, async
     }
 
     res.json(result.data);
-  } catch (error) {
-    res.status(500).json({ error: (error as Error).message });
-  }
-});
-
-router.get("/fetchAirportsFromAvioWiki", verifyUser, async (req, res) => {
-  try {
-    bree.runJob(JobName.ImportAirports);
-
-    res.send("Airport data import started");
   } catch (error) {
     res.status(500).json({ error: (error as Error).message });
   }
